remove: unnecessary queries and run codegen
